+++ src/routes/validationRoutes.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import { validate } from '../controllers/validationController.js';
-import { validateToken } from '../middleware/validation.js';
+import { validateToken } from '../middleware/validationMiddleware.js';
 
 const router = express.Router();
